nanobanana/nanobanana-mcp/pyproject.toml
[project]
name = "nanobanana-mcp"
version = "0.1.0"
description = "MCP server for image generation using Gemini Imagen 3"
dependencies = [
    "mcp[cli]>=1.1.2",
    "python-dotenv>=1.0.1",
    "google-genai>=1.0.0",
    "Pillow>=10.0.0"
]
requires-python = ">=3.10"

[tool.hatch.build.targets.wheel]
packages = ["src/nanobanana_mcp"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
